Remove noise and manipulation of points from the grid class.
By exposing render functions that provide the current point we are able
to have more control without having to bake that into the grid library.

It makes the code far easier to read and manage, and places all the
variables in the project sketch file rather than in the library, meaning
that exploration by tweaking them is easier.